pyproject.toml
[tool.poetry]
name = "afit-staff-management-server"
version = "0.1.0"
description = "This is the API server for my school project, where we are to build a staff management system"
authors = ["emperorsixpacks <andrewdvd10@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
redis = "^5.0.6"
pydantic-settings = "^2.3.3"
pydantic = {extras = ["email"], version = "^2.7.4"}
tortoise-orm = {extras = ["aiosqlite"], version = "^0.21.3"}
aerich = "^0.7.2"
pyyml = "^0.0.2"
passlib = "^1.7.4"
coredis = "^4.17.0"




[tool.aerich]
tortoise_orm = "scripts.orm_config.DEFAULT_CONFIG"
location = "./migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
